Move get_all_subcategories to category.py
