+++ booking/models/category.py
@@ -5,6 +5,18 @@ from django.utils.translation import gettext_lazy as _
 from mptt.models import MPTTModel, TreeForeignKey
 
 
+def get_all_subcategories(categories):
+    """
+    Returns the categories as well as their subcategories.
+    :param ManyRelatedManager categories: categories to get the descendants of.
+    :return list[Category]: list of categories
+    """
+    all_subcategories = list()
+    for category in categories.all():
+        all_subcategories.extend(category.get_descendants(include_self=True))
+    return all_subcategories
+
+
 class Category(MPTTModel):
     name = models.CharField(verbose_name=_("name"), max_length=150, unique=True)
     description = models.CharField(
